+++ forwarder/test/forwarder_test.exs
@@ -11,6 +11,7 @@ defmodule Membrane.RTC.Engine.Endpoint.ForwarderTest do
   alias Membrane.Testing.Pipeline
 
   @stream_id "stream_id"
+  @video_codec :h264
   @forwarder_id {:endpoint, "forwarder"}
 
   setup do
@@ -121,7 +122,8 @@ defmodule Membrane.RTC.Engine.Endpoint.ForwarderTest do
              rtc_engine: self(),
              broadcaster_url: WHIPServer.address(server, @stream_id),
              broadcaster_token: "token",
-             stream_id: @stream_id
+             stream_id: @stream_id,
+             video_codec: @video_codec
            })
          ], group: :forwarder_group, crash_group_mode: :temporary}
     )